move the tests to presto only
